Item693: Try to copy only if it's a file, not a directory
git-svn-id: http://svn.foswiki.org/trunk@1846 0b4bb1d4-4e5a-0410-9cc4-b2b747904278

+++ core/pseudo-install.pl
@@ -165,8 +165,10 @@ sub installModule {
 sub copy_in {
     my ( $moduleDir, $dir, $file ) = @_;
     File::Path::mkpath($dir);
-    File::Copy::copy( "$moduleDir/$file", $file )
-      || die "Couldn't install $file: $!";
+    if( -e "$moduleDir/$file" ) {
+        File::Copy::copy( "$moduleDir/$file", $file )
+        || die "Couldn't install $file: $!";
+    }
     print "Copied $file\n";
 }
